+++ src/check.rs
@@ -473,10 +473,6 @@ impl<'a> VirtualEnv<'a> {
                         }
                     }
                 }
-                Rectify => {
-                    let f = self.pop_func()?;
-                    self.handle_sig(f.signature())?;
-                }
                 SetInverse => {
                     let f = self.pop_func()?;
                     let _inv = self.pop_func()?;